Count the total number of TCP connections
While processing packets we look for unique 4-tuples of: source IP,
destination IP, source port, and destination port; taking into account
that these values may be flipped depending on if the packet is being
sent from client to server or server to client. For each unique tuple we
create a new connection and store its attributes.