Don't use other people's bad code
Although the logic of this function is fine for computing the difference
between two timeval structs, it has the added bonus of changing the two
structs you pass to it, ruining them for future computation. The obvious
fix is storing the values that need to be changed in local variables
before doing so. Passing the structs by value would also work.